uploads feature complete

+++ app/src/main/java/dk/eboks/app/domain/interactors/message/UploadFileInteractor.kt
@@ -15,7 +15,7 @@ interface UploadFileInteractor : Interactor {
 
     interface Output {
         fun onUploadFileComplete()
-        fun onUploadFileProgress(pct : Float)
+        fun onUploadFileProgress(pct : Double)
         fun onUploadFileError(error : ViewError)
     }
 }
